requirements.txt
flask==3.1.0
PyJWT==2.10.1
Flask-Cors==5.0.1
pytz==2025.2
